peeling off movie interface tests
